Amount attribute in manyToMany table between fridge and product was changed from int to double

+++ src/main/java/ntnu/idatt2016/v233/SmartMat/controller/group/FridgeController.java
@@ -67,7 +67,7 @@ public class FridgeController {
 
     @DeleteMapping("/group/delete/product/{fridgeProductId}/{amount}")
     public ResponseEntity<?> deleteAmountFridgeProduct(@PathVariable("fridgeProductId") long fridgeProductId,
-                                                       @PathVariable("amount") int amount){
+                                                       @PathVariable("amount") double amount){
         return fridgeService.deleteAmountFromFridge(fridgeProductId,amount).map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
